Refactor common code into `layout::new::utils`
Also includes some other insignificant refactors
